+++ contactsapp_java/src/main/java/contactsapp/boundary/internal/event/ShouldBeCompany.java
@@ -1,4 +1,4 @@
-package contactsapp.boundary.internal.command_handler;
+package contactsapp.boundary.internal.event;
 
 public class ShouldBeCompany {
 	private String contactId;
