+++ contactsapp_java/src/main/java/contactsapp/boundary/internal/event/ShouldBePerson.java
@@ -1,4 +1,4 @@
-package contactsapp.boundary.internal.command_handler;
+package contactsapp.boundary.internal.event;
 
 public class ShouldBePerson {
 	private String contactId;
