+++ contactsapp_java/src/test/java/contactsapp/boundary/internal/command_handler/HandleEnterEmploymentTest.java
@@ -8,6 +8,8 @@ import org.junit.Test;
 import contactsapp.boundary.internal.domain.ContactList;
 import contactsapp.boundary.internal.event.EmploymentEntered;
 import contactsapp.boundary.internal.event.MissingContact;
+import contactsapp.boundary.internal.event.ShouldBeCompany;
+import contactsapp.boundary.internal.event.ShouldBePerson;
 import contactsapp.command.EnterEmployment;
 
 public class HandleEnterEmploymentTest {
